slight refactor to remove unused code
